feat: create a new community
